+++ premium/backend/src/baserow_premium/api_urls.py
@@ -1,0 +1,9 @@
+from django.conf.urls import url
+
+from baserow_premium.views import UserAdminView
+
+app_name = "baserow_premium"
+
+urlpatterns = [
+    url(r"^users/$", UserAdminView.as_view(), name="users"),
+]