+++ premium/backend/src/baserow_premium/handler.py
@@ -1,0 +1,14 @@
+from django.contrib.auth import get_user_model
+
+User = get_user_model()
+
+
+class PremiumHandler:
+    def get_users(
+        self,
+    ):
+        """"""
+
+        query = User.objects.all()
+
+        return query