+++ premium/web-frontend/modules/baserow_premium/components/UsersAdminContent.vue
@@ -1,24 +1,43 @@
 <template>
-  <div class="user-admin-rows">
-    <div class="user-admin-rows__left">
-      <div class="user-admin-rows__head">
-        <div class="user-admin-rows__field user-admin-rows__field--first">
+  <div class="user-admin-wrapper">
+    <div v-if="!loaded" class="user-admin-rows__initial-loading"></div>
+    <div
+      v-if="loaded"
+      :class="{ 'user-admin-rows__loading': loading }"
+      class="user-admin-wrapper"
+    >
+      <header class="admin_header">
+        <b>User Settings</b>
+        <a
+          ref="contextLink"
+          class="admin-header__link"
+          @click="$refs.context.toggle($refs.contextLink, 'bottom', 'right', 4)"
+        >
+          <i class="fas fa-search"></i>
+        </a>
+      </header>
+      <div class="user-admin-rows">
+        <div class="user-admin-rows__field">
           <div>ID</div>
         </div>
-        <div class="user-admin-rows__field">
+        <div class="user-admin-rows__field user-admin-rows__field-right">
           <i class="fas user-admin-rows__field-icon"></i>
           Username
         </div>
-      </div>
-      <div class="user-admin-rows__body">
+        <div class="user-admin-rows__field">Full Name</div>
+        <div class="user-admin-rows__field">Groups</div>
+        <div class="user-admin-rows__field">Last Login</div>
+        <div class="user-admin-rows__field">Signed Up</div>
+        <div class="user-admin-rows__field">Active</div>
         <div
-          v-for="row in rows"
-          :key="'left-select-row-' + row.id"
-          class="user-admin-rows__row"
+          v-for="cell in cells"
+          :key="'admin-row-' + cell.row_id + '-' + cell.cell_id"
+          :class="{
+            'user-admin-rows__cell-right': cell.cell_id === 2,
+          }"
+          class="user-admin-rows__cell"
         >
-          <div class="user-admin-rows__cell user-admin-rows__cell--first">
-            {{ row.id }}
-          </div>
+          {{ cell.content }}
         </div>
       </div>
       <div class="user-admin-rows__foot">
@@ -30,7 +49,7 @@
               :class="{
                 'user-admin-rows__pagination-button--disabled': page === 1,
               }"
-              @click="fetch(page - 1)"
+              @click="fetchPage(page - 1)"
             >
               <i class="fas fa-caret-left"></i>
             </a>
@@ -38,7 +57,7 @@
               v-model="visiblePage"
               class="input user-admin-rows__pagination-page-input"
               type="number"
-              @keypress.enter="fetch(visiblePage)"
+              @keypress.enter="fetchPage(visiblePage)"
             />
             <div class="user-admin-rows__pagination-count">
               of {{ totalPages }}
@@ -49,7 +68,7 @@
                 'user-admin-rows__pagination-button--disabled':
                   page === totalPages,
               }"
-              @click="fetch(page + 1)"
+              @click="fetchPage(page + 1)"
             >
               <i class="fas fa-caret-right"></i>
             </a>
@@ -57,28 +76,11 @@
         </div>
       </div>
     </div>
-    <div ref="right" class="user-admin-rows__right">
-      <div class="user-admin-rows__head">
-        <div>
-          <i class="fas user-admin-rows__field-icon"></i>
-          Fields
-        </div>
-      </div>
-      <div class="user-admin-rows__body">
-        <div
-          v-for="row in rows"
-          :key="'right-select-row-' + row.id"
-          class="user-admin-rows__row"
-        >
-          Fields
-        </div>
-      </div>
-      <div class="user-admin-rows__foot"></div>
-    </div>
   </div>
 </template>
 
 <script>
+import UserAdminService from '@baserow_premium/services/userAdmin'
 import { notifyIf } from '@baserow/modules/core/utils/error'
 
 export default {
@@ -88,42 +90,18 @@ export default {
     return {
       loading: false,
       loaded: false,
-      rows: [{ name: 'test', id: 1 }],
       search: '',
       visibleSearch: '',
       page: 1,
       visiblePage: 1,
       totalPages: null,
+      cells: [],
     }
   },
+  async fetch() {
+    await this.fetchPage(1)
+  },
   methods: {
-    /**
-     * Returns the scrollable element for the scrollbar.
-     */
-    /**
-     * Fetches all the fields of the given table id. We need the fields so that we can
-     * show the data in the correct format.
-     */
-    async fetchFields(tableId) {
-      try {
-        await true
-        // const { data } = await FieldService(this.$client).fetchAll(tableId)
-        // data.forEach((part, index, d) => {
-        //   populateField(data[index], this.$registry)
-        // })
-        // const primaryIndex = data.findIndex((item) => item.primary === true)
-        // this.primary =
-        //   primaryIndex !== -1 ? data.splice(primaryIndex, 1)[0] : null
-        // this.fields = data
-      } catch (error) {
-        this.loading = false
-        notifyIf(error, 'row')
-      }
-    },
-    /**
-     * Does a row search in the table related to the state. It will also reset the
-     * pagination.
-     */
     async doSearch(query) {
       this.search = query
       this.totalPages = null
@@ -133,10 +111,56 @@ export default {
      * Fetches the rows of a given page and adds them to the state. If a search query
      * has been stored in the state then that will be remembered.
      */
-    async fetch(page) {},
-    /**
-     * Called when the user selects a row.
-     */
+    async fetchPage(page) {
+      if (
+        this.totalPages !== null &&
+        this.totalPages !== 0 &&
+        (page > this.totalPages || page < 1)
+      ) {
+        console.log('SOME DUBM SHIT')
+        this.visiblePage = this.page
+        return
+      }
+
+      this.loading = true
+      this.loaded = false
+
+      try {
+        const { data: userData } = await UserAdminService(
+          this.$client
+        ).fetchPage(page)
+        const newCells = []
+        for (const user of userData.results) {
+          const attrs = [
+            user.id,
+            user.username,
+            user.full_name,
+            user.groups.map((e) => e.name).join(', '),
+            user.last_login,
+            user.date_joined,
+            user.is_active,
+          ]
+          let i = 1
+          for (const attr of attrs) {
+            newCells.push({
+              row_id: user.id,
+              cell_id: i++,
+              content: attr,
+            })
+          }
+        }
+        this.page = page
+        this.visiblePage = page
+        this.totalPages = Math.ceil(userData.count / 100)
+        this.loaded = true
+        this.cells = newCells
+      } catch (error) {
+        notifyIf(error, 'row')
+        this.loaded = false
+      }
+
+      this.loading = false
+    },
   },
 }
 </script>