+++ premium/web-frontend/modules/baserow_premium/services/userAdmin.js
@@ -1,0 +1,7 @@
+export default (client) => {
+  return {
+    fetchPage(page) {
+      return client.get(`/premium/users/?page=${page}`)
+    },
+  }
+}